Notification should not hide content according to ux
Issue description:
Notification will hide some content when it popup.

According to ux requirement, we should make it not extend to the whole widh of UI, and should not hide the content.

Fix:
Modify the position from 'absolute' to 'relative', so that it will not hide the content.

Antilles-original-commit: 2240693a1d6d76238964509b4d4283258657089e
Antilles-max-blob-size: 26214400
Antilles-scrubbed-blob: 9a1a74e921daddc206a912ea066958667e91bd48 src/ui/harbor_ui (33925696 bytes)