Update CFG migrator to match the new redis setting (#5567)
There's a limitation in the new harbor.cfg file the redis url will be
hard coded to the internal redis instance.  We expect user to manually
edit the .cfg file if he's using a customized redis, this would be a
very rare case.  We'll mention it in the release note of 1.6

Antilles-original-commit: 3dc95ee7e28aad6326435a48de7e6e4b10310b6a
Antilles-max-blob-size: 26214400