Update automation test case "Project Level Image Serverity Policy" (#6037)
Currently, the case "Project Level Image Serverity Policy" uses the project library as the target project, but the project library has been set to scan image automatically when images are pushing, this will cause case failure. This commit creates a new project in the case to aviod the failure.

Signed-off-by: Wenkai Yin <yinw@vmware.com>
Antilles-original-commit: 16c9ffad2ce5e6124bb735883444198636fa7d68
Antilles-max-blob-size: 26214400