add new api tests (#6127)
Add API test libraries which were based on swagger doc,
it's including user and project , also add one scenario
test case which is test_add_member_to_private_project to
verify a user can see a project when the uesr is the member of it.
Delete 2 old test case.Delete a quota.

Signed-off-by: danfengliu <danfengl@vmware.com>
Antilles-original-commit: 894e0591bb6b7a9969d0ba5e11a6146ba2eb778d
Antilles-max-blob-size: 26214400