Wait for manifest in notification handler
There's an issue in registry 2.6.x, that when the webhook is sent the
manifest of the image may not be written.
For details: https://github.com/docker/distribution/issues/2625

This will cause issue in "scan on push" or replication.
This commit mitigates the issue by adding retries in notification
handler.

Signed-off-by: Daniel Jiang <jiangd@vmware.com>
Antilles-original-commit: abe728325bbff6bf5763a2991e49bbb54d4d9fee
Antilles-max-blob-size: 26214400