Fix Swagger client generation for a few chartrepo endpoints
It looks like the Swagger for `ChartVersions` and `ChartVersionDetails`
were missing fields to properly generate a client.

This fixes the problem by adding the correct `refs` so that the client
produces an output when making an API call.

In Go, this generates something like:

```Go
/*GetChartrepoRepoChartsNameVersionOK handles this case with default header values.

Successfully retrieved the chart version
*/
type GetChartrepoRepoChartsNameVersionOK struct {
	Payload *models.ChartVersionDetails
}

...

/*GetChartrepoRepoChartsNameOK handles this case with default header values.

Retrieved all versions of the specified chart
*/
type GetChartrepoRepoChartsNameOK struct {
	Payload models.ChartVersions
}
```

Signed-off-by: Dan Norris <dan.norris@netapp.com>
Antilles-original-commit: fa4280887d3b7a0def0b722c01bfdd6f4df06148
Antilles-max-blob-size: 26214400