add internal reg request handler chain
this is for internal registry api call, the request should be intercpeted by quota middlerwares, like retag and delete.
Note: The api developer has to know that if the internal registry call in your api, please consider to use
NewRepositoryClientForLocal() to init the repository client, which can handle quota change.

Signed-off-by: wang yan <wangyan@vmware.com>
Antilles-original-commit: 4410cc93f90e3f3598ff9b2c9dbc55cd2d0d9347
Antilles-max-blob-size: 26214400