Add merge slash middleware
There are code in the core component to conditionally execute code based
on the pattern of url path, and different ingress controller or reverse
proxy may handle the dup slashes in the url path differently.
This commit merge dup slashes in the url paths to make things more
consistent.

Signed-off-by: Daniel Jiang <jiangd@vmware.com>
Antilles-original-commit: f4ac81b710e8622ca8dc58f1296c180e780924fb
Antilles-max-blob-size: 26214400