README.md: fix broken links to badges and Swagger editor (#16910)
* README.md: fix links to badges

the links to badges are still referring to "master" branch instead of "main". this commit fixes it.

* README.md: fix Restful API links to swagger editor

The links to swagger editor had the URL paths to the master branch which has hence been renamed. This leads to swagger editor failing to open the API links. this commit fixes that.

Antilles-original-commit: 5afd11015a1390d0b0dc6c3d7323cb9a90c5f3d7
Antilles-max-blob-size: 26214400