Generate URI of token service via Host in request (#21898)
This commit update the flow to generate URL of token service, which will
first try to use the Host in request.  This will help the situation when
Harbor is configured to serve via a hostname but some client needs to
pull artifacts from Harbor via IP due to limitations in the environment.

Signed-off-by: Daniel Jiang <daniel.jiang@broadcom.com>
Antilles-original-commit: 70306dca0cefbb17ecb7a6810eb28f84f28ecb63
Antilles-max-blob-size: 26214400