udpate storage to s3 (#21999)
move the build storage from google storage to the CNCF S3 storage

Currently, we use the internal GCR to store all dev builds for nightly testing, development, and as candidates for RC and GA releases. However, this internal Google storage will no longer be available, this pull request it to move to the CNCF-hosted S3 storage.

Signed-off-by: wang yan <wangyan@vmware.com>
Antilles-original-commit: 2faff8e6af98730c32d5f82f1dcdece472ac73f4
Antilles-max-blob-size: 26214400